MACREADY           Decanite...?  Thermite charges...?
The tape jump cuts again showing a long shot of the
markings.  No Norwegian in sight.  An explosion kicks up
the ice.  A beat as the ice sprays to the ground.  Then
the camera appears to shake as the ground beneath it
quivers.

Another immense explosion follows.  An earthquake-like
force throws the camera to the ground.